update parse function and move internal widget types file
